Convert to markdown
For an open source methodology, it was surprisingly difficult to find an
'open' copy of the original document. I've decided to fork the github
repo of somebody else who was nice enough to go through that effort and
publish the document somewhere accessible.

I thought it would ne nice to "modify" and create a derivative for
people like me who find pdf slide shows with tons of background images
difficult to read from an accessibility perspective. For now I've just
copied the text and attempted to capture the same spirit of the original
formatting. Next I'd like to reformat it as an epub so that it will be
easier to read on my e-reader.